refactor(runtime): split promise/then.rs and proxy.rs under the 2000-line cap
Pure relocation, no functional changes:
- promise/then.rs (2135) -> promise/reactions.rs (298): the settle-listener
  and overflow-reaction side tables with their attach/take/enqueue helpers,
  GC scanners, dead-promise death hooks, and copied-minor cleanup. then.rs
  is now 1851 lines. Re-exports wired through promise/mod.rs
  (js_promise_attach_settle_listener) and scanners.rs/mod.rs callers
  repointed to super::reactions.
- proxy.rs (2052) -> proxy/apply_construct.rs (242): the [[Call]]/[[Construct]]
  exotic behavior (js_proxy_apply/js_proxy_construct, forward_apply/
  forward_construct) plus the is_callable_function/is_constructor_function
  classifiers used by the Reflect trap validators. proxy.rs is now 1823
  lines. Same recipe as the existing proxy/metadata.rs split.
